migrate all shared code out of the transformer files, and remove Transform arguments
